requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
ale-py=0.8.1=pypi_0
appdirs=1.4.4=pypi_0
attrs=23.1.0=pypi_0
box2d-py=2.3.5=pypi_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.0.1=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.1=pypi_0
contourpy=1.0.7=pypi_0
cv=1.0.0=pypi_0
cycler=0.11.0=pypi_0
cython=0.29.34=pypi_0
decorator=4.4.2=pypi_0
docker-pycreds=0.4.0=pypi_0
et-xmlfile=1.1.0=pypi_0
farama-notifications=0.0.4=pypi_0
fasteners=0.18=pypi_0
fonttools=4.38.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glfw=2.5.9=pypi_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.51.1=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
gymnasium=0.28.1=pypi_0
idna=3.4=pypi_0
imageio=2.25.1=pypi_0
imageio-ffmpeg=0.4.8=pypi_0
importlib-metadata=6.0.0=pypi_0
importlib-resources=5.10.2=pypi_0
iniconfig=2.0.0=pypi_0
isaacgym=1.0rc4=dev_0
jax-jumpy=1.0.0=pypi_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
lz4=4.3.2=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.0=pypi_0
moviepy=1.0.3=pypi_0
mujoco=2.2.0=pypi_0
mujoco-py=2.1.2.14=pypi_0
ncurses=6.4=h6a678d5_0
ninja=1.11.1=pypi_0
numpy=1.24.2=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.0.74=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py38h06a4308_0
pluggy=1.2.0=pypi_0
proglog=0.1.10=pypi_0
protobuf=4.21.12=pypi_0
psutil=5.9.4=pypi_0
py=1.11.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pybullet=3.2.5=pypi_0
pycparser=2.21=pypi_0
pygame=2.1.0=pypi_0
pyopengl=3.1.7=pypi_0
pyparsing=3.0.9=pypi_0
pytest=7.0.1=pypi_0
python=3.8.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.2.2=pypi_0
scipy=1.10.0=pypi_0
sentry-sdk=1.15.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.40.1=h5082296_0
swig=4.1.1=pypi_0
tensorboard=2.12.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=1.13.1=pypi_0
torchvision=0.14.1=pypi_0
tqdm=4.64.1=pypi_0
typing-extensions=4.5.0=pypi_0
urllib3=1.26.14=pypi_0
wandb=0.15.8=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
zipp=3.13.0=pypi_0
zlib=1.2.13=h5eee18b_0
